feat: controllerにエラーハンドリングを追加する

+++ packages/server/src/todos/controller.ts
@@ -3,6 +3,7 @@ import {
   createTodo,
   deleteTodo,
   findAllTodos,
+  isExistTodo,
   setCompleteTodo,
   setIncompleteTodo,
 } from "./model";
@@ -26,18 +27,42 @@ router.post("/todos", async (req, res) => {
 
 router.delete("/todos/:id", async (req, res) => {
   const { id } = req.params;
+
+  const existTodo = await isExistTodo(id);
+
+  if (!existTodo) {
+    res.status(404).end();
+    return;
+  }
+
   await deleteTodo(id);
   res.status(200).end();
 });
 
 router.put("/todos/:id/completed", async (req, res) => {
   const { id } = req.params;
+
+  const existTodo = await isExistTodo(id);
+
+  if (!existTodo) {
+    res.status(404).end();
+    return;
+  }
+
   await setCompleteTodo(id);
   res.status(204).end();
 });
 
 router.put("/todos/:id/uncompleted", async (req, res) => {
   const { id } = req.params;
+
+  const existTodo = await isExistTodo(id);
+
+  if (!existTodo) {
+    res.status(404).end();
+    return;
+  }
+
   await setIncompleteTodo(id);
   res.status(204).end();
 });
